chore: preserve binaries across reboots

+++ docker/all-in-one/init/configure-adminapi.sh
@@ -5,17 +5,30 @@ ADMIN_API_CONF=/etc/adminapi/adminapi.yaml
 touch /var/log/services/adminapi.log
 
 ADMINAPI_CUSTOM_DIR="${DATA_VOLUME_MOUNTPOINT}/etc/adminapi"
-mkdir -p "${ADMINAPI_CUSTOM_DIR}"
-if [ ! -f "${CONFIGURED_FLAG_PATH}" ]; then
-  echo "Copying existing custom adminapi config from /etc/adminapi to ${ADMINAPI_CUSTOM_DIR}"
-  cp -R "/etc/adminapi/." "${ADMINAPI_CUSTOM_DIR}/"
-fi
 
-rm -rf "/etc/adminapi"
-ln -s "${ADMINAPI_CUSTOM_DIR}" "/etc/adminapi"
-chown -R adminapi:adminapi "/etc/adminapi"
-chown -R adminapi:adminapi "${ADMINAPI_CUSTOM_DIR}"
-chmod g+rx "${ADMINAPI_CUSTOM_DIR}"
+if [ "${DATA_VOLUME_MOUNTPOINT}" ]; then
+  mkdir -p "${ADMINAPI_CUSTOM_DIR}"
+  if [ ! -f "${CONFIGURED_FLAG_PATH}" ]; then
+    echo "Copying existing custom adminapi config from /etc/adminapi to ${ADMINAPI_CUSTOM_DIR}"
+    cp -R "/etc/adminapi/." "${ADMINAPI_CUSTOM_DIR}/"
+  fi
+
+  if [ ! -f "${ADMINAPI_CUSTOM_DIR}/supabase-admin-api" ]; then
+    echo "Copying existing adminapi binary from /opt/supabase-admin-api to ${ADMINAPI_CUSTOM_DIR}"
+    cp "/opt/supabase-admin-api" "${ADMINAPI_CUSTOM_DIR}/supabase-admin-api"
+  fi
+
+  rm -rf "/etc/adminapi"
+  ln -s "${ADMINAPI_CUSTOM_DIR}" "/etc/adminapi"
+  chown -R adminapi:adminapi "/etc/adminapi"
+
+  rm -f "/opt/supabase-admin-api"
+  ln -s "${ADMINAPI_CUSTOM_DIR}/supabase-admin-api" "/opt/supabase-admin-api"
+  chown -R adminapi:adminapi "/opt/supabase-admin-api"
+
+  chown -R adminapi:adminapi "${ADMINAPI_CUSTOM_DIR}"
+  chmod g+rx "${ADMINAPI_CUSTOM_DIR}"
+fi
 
 if [ -f "${INIT_PAYLOAD_PATH:-}" ]; then
   echo "init adminapi payload"
